+++ src/utils/schema.ts
@@ -1,10 +1,4 @@
-import {
-    BuildArgProxy,
-    DerivedValueProxy,
-    Ref,
-    TypeFactory,
-} from '../type-factory';
-import { ERROR_MESSAGES } from '../constants';
+import { DerivedValueProxy, Ref, TypeFactory } from '../type-factory';
 import { FactorySchema } from '../types';
 import { isOfType, isRecord } from './guards';
 import { throwIfPromise } from './general';
@@ -69,42 +63,4 @@ export async function parseFactorySchemaAsync<T>(
     }
 
     return output as T;
-}
-
-export function validateFactorySchema<T extends FactorySchema<any>>(
-    schema: T,
-): T {
-    const missingValues: string[] = [];
-    Object.entries(schema).forEach(([key, value]) => {
-        if (value instanceof BuildArgProxy) {
-            missingValues.push(key);
-        }
-    });
-    if (missingValues.length) {
-        throw new Error(
-            ERROR_MESSAGES.MISSING_BUILD_ARGS.replace(
-                ':missingArgs',
-                missingValues.join(', '),
-            ),
-        );
-    }
-    return schema;
-}
-
-export function validateFactoryResult<T>(factoryResult: T): T {
-    const missingValues: string[] = [];
-    Object.entries(factoryResult).forEach(([key, value]) => {
-        if (value instanceof DerivedValueProxy) {
-            missingValues.push(key);
-        }
-    });
-    if (missingValues.length) {
-        throw new Error(
-            ERROR_MESSAGES.MISSING_DERIVED_PARAMETERS.replace(
-                ':missingValues',
-                missingValues.join(', '),
-            ),
-        );
-    }
-    return factoryResult;
 }
