--- tests/utils.ts
@@ -1,30 +1,0 @@
-import { ComplexObject } from './test-types';
-
-export const defaults: ComplexObject = {
-    name: 'testObject',
-    value: null,
-};
-
-export const twoLevelDefaults: ComplexObject = {
-    ...defaults,
-    options: {
-        type: '1',
-    },
-};
-
-export const threeLevelDefaults: ComplexObject = {
-    ...twoLevelDefaults,
-    options: {
-        type: '1',
-        children: [defaults],
-    },
-};
-
-export const annoyinglyComplexObject: any = {
-    ...twoLevelDefaults,
-    options: {
-        type: '1',
-        children: [threeLevelDefaults],
-        arrayOfArray: [[[threeLevelDefaults]]],
-    },
-};
