+++ tests/utils/general.spec.ts
@@ -1,0 +1,11 @@
+import { ERROR_MESSAGES } from '../../src';
+import { throwIfPromise } from '../../src/utils/general';
+
+describe('throwIfPromise', () => {
+    it('throws when promise and passes value otherwise', () => {
+        expect(() =>
+            throwIfPromise(new Promise((resolve) => resolve(null)), 'test'),
+        ).toThrow(ERROR_MESSAGES.PROMISE_VALUE.replace(':key', 'test'));
+        expect(throwIfPromise({}, 'test')).toEqual({});
+    });
+});
