+++ tests/utils/guards.spec.ts
@@ -1,0 +1,24 @@
+import { isPromise, isRecord } from '../../src/utils/guards';
+
+describe('isRecord', () => {
+    it('returns true for records and false for non-records', () => {
+        expect(isRecord({})).toBeTruthy();
+        expect(isRecord(new Map())).toBeTruthy();
+        expect(isRecord(new WeakMap())).toBeTruthy();
+        expect(isRecord(new Map())).toBeTruthy();
+        expect(isRecord([])).toBeFalsy();
+        expect(isRecord(new Set())).toBeFalsy();
+        expect(isRecord(() => null)).toBeFalsy();
+        expect(isRecord('')).toBeFalsy();
+        expect(isRecord(1)).toBeFalsy();
+    });
+});
+
+describe('isPromise', () => {
+    it('returns true for promises and false for non promises', () => {
+        expect(isPromise(new Promise((resolve) => resolve(null)))).toBeTruthy();
+        expect(isPromise({})).toBeFalsy();
+        expect(isPromise(null)).toBeFalsy();
+        expect(isPromise(1)).toBeFalsy();
+    });
+});
