+++ tests/utils/options.spec.ts
@@ -1,0 +1,47 @@
+import { parseOptions } from '../../src/utils/options';
+
+describe('parseOptions', () => {
+    it('returns [undefined, undefined] when passed undefined', () => {
+        expect(parseOptions(undefined, 1)).toEqual([undefined, undefined]);
+    });
+    it('returns [options, undefined] when passed options object', () => {
+        const options = { key: 'value' };
+        expect(parseOptions(options, 1)).toEqual([options, undefined]);
+    });
+    it('calls options function with iteration', () => {
+        const options = jest.fn((iteration: number) => ({ key: iteration }));
+        const result = parseOptions(options, 1);
+        expect(options).toHaveBeenCalledWith(1);
+        expect(result).toEqual([{ key: 1 }, undefined]);
+    });
+    describe('handles OverridesAndFactory object correctly', () => {
+        it('returns both overrides and factory when passed both', () => {
+            const overrides = { key: 'value' };
+            const factory = jest.fn();
+            expect(parseOptions({ overrides, factory }, 1)).toEqual([
+                overrides,
+                factory,
+            ]);
+        });
+        it('returns [undefined, factory] when only factory is passed', () => {
+            const factory = jest.fn();
+            expect(parseOptions({ factory }, 1)).toEqual([undefined, factory]);
+        });
+        it('returns [overrides, undefined] when only overrides are passed', () => {
+            const overrides = { key: 'value' };
+            expect(parseOptions({ overrides }, 1)).toEqual([
+                overrides,
+                undefined,
+            ]);
+        });
+        it('calls function overrides with iteration', () => {
+            const overrides = jest.fn((iteration: number) => ({
+                key: iteration,
+            }));
+            const factory = jest.fn();
+            const result = parseOptions({ overrides, factory }, 1);
+            expect(overrides).toHaveBeenCalledWith(1);
+            expect(result).toEqual([{ key: 1 }, factory]);
+        });
+    });
+});
